+++ src/CodeIndex/Program.cs
@@ -53,6 +53,7 @@ return args[0] switch
     "inspect" => QueryCommandRunner.RunInspect(args[1..], jsonOptions),
     "outline" => QueryCommandRunner.RunOutline(args[1..], jsonOptions),
     "status" => QueryCommandRunner.RunStatus(args[1..], jsonOptions),
+    "languages" => QueryCommandRunner.RunLanguages(args[1..], jsonOptions),
     "index" => IndexCommandRunner.Run(args[1..], jsonOptions),
     _ when IsProjectPathArg(args[0])
         => IndexCommandRunner.Run(args, jsonOptions),